frontend/requirements.txt
ultralytics==8.3.50
streamlit
opencv-python
numpy
Pillow
